requirements.txt
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
azure-appconfiguration==1.6.0; python_version >= '3.8'
azure-appconfiguration-provider==1.2.0; python_version >= '3.6'
azure-core==1.30.2; python_version >= '3.8'
azure-keyvault-secrets==4.8.0; python_version >= '3.8'
azure-storage-blob==12.20.0; python_version >= '3.8'
cachecontrol==0.14.0; python_version >= '3.7'
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.8; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.1.1; python_version >= '3.8'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
firebase-admin==6.5.0; python_version >= '3.7'
geoalchemy2==0.15.2; python_version >= '3.7'
geographiclib==2.0; python_version >= '3.7'
geojson==3.1.0; python_version >= '3.7'
geopy==2.4.1; python_version >= '3.7'
google-api-core[grpc]==2.19.0; python_version >= '3.7'
google-api-python-client==2.133.0; python_version >= '3.7'
google-auth==2.30.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-firestore==2.16.0; python_version >= '3.7'
google-cloud-storage==2.17.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.7.1; python_version >= '3.7'
googleapis-common-protos==1.63.1; python_version >= '3.7'
googlemaps==4.10.0; python_version >= '3.5'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.64.1; python_version >= '3.8'
grpcio-status==1.62.2; python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
h2==4.1.0; python_full_version >= '3.6.1'
hpack==4.0.0; python_full_version >= '3.6.1'
httpcore==1.0.5; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httptools==0.6.1; python_full_version >= '3.8.0'
httpx==0.27.0; python_version >= '3.8'
hyperframe==6.0.1; python_full_version >= '3.6.1'
idna==3.7; python_version >= '3.5'
isodate==0.6.1
jinja2==3.1.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
methoddispatch==3.0.2
msgpack==1.0.8; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
openai==1.35.13; python_full_version >= '3.7.1'
orjson==3.10.5; python_version >= '3.8'
overpass==0.7
packaging==24.1; python_version >= '3.8'
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.3; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.4; python_version >= '3.8'
pydantic-core==2.18.4; python_version >= '3.8'
pyee==11.1.0; python_version >= '3.8'
pygeohash==1.2.0
pygments==2.18.0; python_version >= '3.8'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pymongo==4.8.0; python_version >= '3.8'
pyparsing==3.1.2; python_full_version >= '3.6.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
schedule==1.2.2; python_version >= '3.7'
shapely==2.0.4; python_version >= '3.7'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.32; python_version >= '3.7'
sqlmodel==0.0.21; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
ujson==5.10.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.1; python_version >= '3.8'
uvicorn[standard]==0.30.1; python_version >= '3.8'
uvloop==0.19.0; python_full_version >= '3.8.0'
watchdog==4.0.1; python_version >= '3.8'
watchfiles==0.22.0; python_version >= '3.8'
websockets==12.0; python_version >= '3.8'
